refactor(numeric): flip number_to_string mirrors + C-ABI rows to int (#611)
Flips the four `number_to_string` mirror functions in `main.sfn`,
`emit_native_state.sfn`, `llvm/utils.sfn`, and
`llvm/expression_lowering/native/core_text.sfn` from `: number` to
`: int`, including their body locals and the two `powers: number[]`
constant tables (now `: int[]`). Removes the `// alias-coverage:
seed-ABI constraint` markers and the stale "seed alpha.18 carve-out"
banner in `main.sfn`.

Flips the C-ABI rows `sailfin_cli_main` and `native_cli_main` in
`cli_main.sfn` from `-> number` to `-> int`, with a paired one-line
update to `runtime/native/src/native_driver.c` so the C-side declare
of `sailfin_cli_main__cli_main` now reads `int64_t` (and the `rc`
local in `main()` is `int64_t` to match).

Discovered scope addition: `_trj_per_test_duration` in
`cli_commands.sfn` was declared `-> float` despite always returning
the `int` `quotient`. Its result was passed straight into
`render_test_event(duration_ms: int, ...)`, producing the one
remaining pre-existing `ABI primitive mismatch` warning surfaced by
`make compile` after the mirror flip cleared the seed-ABI sites.
Returning `int` instead of `float` is the same shape of fix the
issue describes, and is required to satisfy the acceptance criterion
of zero `ABI primitive mismatch` warnings.

Updates rows 115/116 of `docs/proposals/slice-e3a-semantic-audit.md`
to reflect the new `: int` shape and the paired C-driver flip.

Closes #611

+++ runtime/native/src/native_driver.c
@@ -30,7 +30,7 @@ extern char *compile_to_llvm(char *source);
 // NOTE: `native_cli_main` has historically varied in symbol spelling across
 // compiler versions (plain vs module-suffixed). Call the stable mangled symbol
 // for the underlying CLI implementation instead.
-extern double sailfin_cli_main__cli_main(SailfinPtrArray *argv);
+extern int64_t sailfin_cli_main__cli_main(SailfinPtrArray *argv);
 
 // Optional process-exit arena telemetry. Gated on SAILFIN_DUMP_ARENA_STATS=1
 // alone — when set, the atexit hook always runs.  At exit it checks
@@ -590,7 +590,7 @@ int main(int argc, char **argv)
 
             _trace_ptr_array("native-cli", &args);
 
-            double rc = sailfin_cli_main__cli_main(&args);
+            int64_t rc = sailfin_cli_main__cli_main(&args);
             free(argv_copy);
             free(runtime_root);
             free(binary_dir);
